use long for expiry time in milliseconds

+++ src/main/java/app/attestation/server/Maintenance.java
@@ -11,7 +11,7 @@ import java.util.logging.Logger;
 
 class Maintenance implements Runnable {
     private static final long WAIT_MS = 24 * 60 * 60 * 1000;
-    private static final int DELETE_EXPIRY_MS = 7 * 24 * 60 * 60 * 1000;
+    private static final long DELETE_EXPIRY_MS = 7L * 24 * 60 * 60 * 1000;
     private static final int KEEP_BACKUPS = 28;
 
     private static final Logger logger = Logger.getLogger(Maintenance.class.getName());
